chore: apply the recent section as a component
